🚚 Moved techIcon to shared components
