+++ components/myExperience/MyExperience.tsx
@@ -2,7 +2,7 @@ import styles from './myExperience.module.scss'
 import { useEffect, useState } from 'react';
 import Experience from '@/interfaces/experience';
 import { tldr } from '@/data/tldrExperience';
-import TechIcon from '../experienceCard/TechIcon';
+import TechIcon from '../shared/techIcon/TechIcon';
 import Link from 'next/link';
 
 function MyExperience() {
